add route to tupaia server
