+++ packages/lesmis-server/src/routes/LogoutRoute.ts
@@ -3,7 +3,7 @@
  * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
  */
 
-import { UnauthenticatedRoute } from './UnauthenticatedRoute';
+import { UnauthenticatedRoute } from '@tupaia/tupaia-server';
 
 export class LogoutRoute extends UnauthenticatedRoute {
   async buildResponse() {
