+++ packages/lesmis-server/src/routes/Route.ts
@@ -2,44 +2,17 @@
  * Tupaia
  * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
  */
-import { NextFunction, Response } from 'express';
-import { respond, PermissionsError, UnauthenticatedError } from '@tupaia/utils';
+import { PermissionsError } from '@tupaia/utils';
+import { Route as BaseRoute } from '@tupaia/tupaia-server';
 
-import { LesmisRequest, LesmisResponseBody, SessionCookie } from '../types';
-import { LesmisSessionModel, LesmisSessionType } from '../models';
+import { LesmisSessionType } from '../models';
 import { MeditrakConnection, ReportConnection } from '../connections';
 import { LESMIS_PERMISSION_GROUP } from '../constants';
 
-export class Route {
-  req: LesmisRequest;
-
-  res: Response;
-
-  next: NextFunction;
-
-  sessionModel: LesmisSessionModel;
-
-  sessionCookie?: SessionCookie;
-
-  session!: LesmisSessionType;
-
+export class Route extends BaseRoute {
   meditrakConnection?: MeditrakConnection;
 
   reportConnection?: ReportConnection;
-
-  constructor(req: LesmisRequest, res: Response, next: NextFunction) {
-    this.req = req;
-    this.res = res;
-    this.next = next;
-
-    const { sessionModel, sessionCookie } = req;
-    this.sessionModel = sessionModel;
-    this.sessionCookie = sessionCookie;
-  }
-
-  respond(responseBody: LesmisResponseBody, statusCode: number) {
-    respond(this.res, responseBody, statusCode);
-  }
 
   async handle() {
     // All routes will be wrapped with an error catcher that simply passes the error to the next()
@@ -56,20 +29,6 @@ export class Route {
     }
   }
 
-  async getSession() {
-    const sessionId = this.sessionCookie?.id;
-    if (!sessionId) {
-      throw new UnauthenticatedError('User not authenticated');
-    }
-
-    const session: LesmisSessionType = await this.sessionModel.findById(sessionId);
-    if (!session) {
-      throw new UnauthenticatedError('Session not found in database');
-    }
-
-    return session;
-  }
-
   async verifyAuth(): Promise<LesmisSessionType> {
     const session = await this.getSession();
     const { accessPolicy } = session;
@@ -79,9 +38,5 @@ export class Route {
     }
 
     return session;
-  }
-
-  async buildResponse(): Promise<Record<string, unknown>> {
-    throw new Error('Any Route must implement "buildResponse"');
   }
 }
