+++ packages/lesmis-server/src/routes/TestRoute.ts
@@ -3,7 +3,7 @@
  * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
  */
 
-import { UnauthenticatedRoute } from './UnauthenticatedRoute';
+import { UnauthenticatedRoute } from '@tupaia/tupaia-server';
 
 export class TestRoute extends UnauthenticatedRoute {
   async buildResponse() {
