+++ packages/tupaia-server/src/UnauthenticatedRoute.ts
@@ -1,6 +1,7 @@
-/**
+/*
  * Tupaia
  * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
+ *
  */
 
 import { Route } from './Route';
